pyproject.toml
[project]
name = "rag-starter"
version = "0.11.0"
description = "General-purpose RAG starter"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "base32-crockford>=0.3.0",
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.4.1",
    "datasets>=4.5.0",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.9",
    "langchain-text-splitters>=0.4.0",
    "lxml>=6.0.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=2.2.0",
    "psutil>=7.2.2",
    "pypdf>=6.6.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "ragas>=0.4.3",
    "requests>=2.32.5",
    "rich>=14.3.2",
    "ruff>=0.15.0",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.43",
]

[dependency-groups]
dev = [
    "commitizen>=4.13.2",
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true
